Update Main.java
Unable to determine how to enable user input as code is skewing to input only from the worker at the establishment.( or worker keys in customer/client info). 

Additional research required for development of application to streamline process of input. (i.e.  customers should be able to input from web or app to make this project relevant)